Nov week1

+++ header.h
@@ -27,7 +27,7 @@ using std::vector;
 template<class T>
 void print_vector(vector<T>& v) {
     cout << "[";
-    for (int i = 0; i < v.size() - 1; i++) {
+    for (int i = 0; i < (int)v.size() - 1; i++) {
         cout << v[i] << ",";
     }
     if (v.size() > 0) {
@@ -41,7 +41,7 @@ void print_vector(vector<vector<T>>& v) {
     cout << "[";
     for(int i = 0; i < v.size(); i++) {
         cout << "[";
-        for(int j = 0; j < v[i].size() - 1; j++) {
+        for(int j = 0; j < (int)v[i].size() - 1; j++) {
             cout << v[i][j] << ",";
         }
         cout << v[i][v[i].size() - 1];
